Transitioned the idea of changing a brush stroke to another...sort of
The message of drawing ellipses when using a different brush stroke was achived but since these messages are being sent and recieved, the same ellipses are also being drawn along with the line function I set up in intBrush(). To be honest I find this kind of cool. It's like everyone has their own trail or map on their phone or laptop but the ambiguity is captured on the screen that is not being drawn on.

Thinking of adding geolocation so instead of mousedragging the location of the user's phone will make traces...